refactor: Delete modal [#132]
