+++ src/lib/components/collections/collections-list/FileExplorer.svelte
@@ -15,7 +15,6 @@
   import Spinner from "$lib/components/Transition/Spinner.svelte";
   import threedotIcon from "$lib/assets/3dot.svg";
   import { CollectionService } from "$lib/services/collection.service";
-  import FolderPopup from "$lib/components/Modal/FolderPopup.svelte";
   import { selectMethodsStore } from "$lib/store/methods";
   import { onDestroy } from "svelte";
   import { generateSampleFolder } from "$lib/utils/sample/folder.sample";
@@ -26,6 +25,9 @@
   import angleRight from "$lib/assets/angleRight.svg";
   import MixpanelEvent from "$lib/utils/mixpanel/MixpanelEvent";
   import { Events } from "$lib/utils/enums/mixpanel-events.enum";
+  import ModalWrapperV1 from "$lib/components/Modal/ModalWrapperV1.svelte";
+  import CustomButton from "$lib/components/buttons/CustomButton.svelte";
+  import { notifications } from "$lib/utils/notifications";
 
   let expand: boolean = false;
   export let explorer;
@@ -232,25 +234,104 @@
 
   let workspaceId = currentWorkspaceId;
 
+  let requestCount: number;
+  let requestIds = [];
   $: {
     if (activePath) {
       if (activePath.folderId === explorer.id) {
         expand = true;
       }
     }
+    if (explorer) {
+      requestIds.length = 0;
+      requestCount = 0;
+      requestCount = explorer?.items?.length;
+      if (explorer?.items) {
+        requestIds = explorer?.items?.map((element) => {
+          return element.id;
+        });
+      }
+      requestIds.push(explorer?.id);
+    }
   }
+
+  // Delete folder functions
+  let deleteLoader: boolean = false;
+
+  const handleDelete = async () => {
+    deleteLoader = true;
+    const response = await collectionService.deleteFolderInCollection(
+      workspaceId,
+      collectionId,
+      explorer.id,
+    );
+
+    if (response.isSuccessful) {
+      collectionsMethods.deleteRequestOrFolderInCollection(
+        collectionId,
+        explorer.id,
+      );
+
+      notifications.success(`"${explorer.name}" Folder deleted.`);
+      deleteLoader = false;
+      collectionsMethods.removeMultipleTabs(requestIds);
+      handleFolderPopUp(false);
+    } else {
+      notifications.error("Failed to delete the Folder.");
+      deleteLoader = false;
+    }
+  };
 </script>
 
-{#if isFolderPopup}
-  <FolderPopup
-    {collectionsMethods}
-    {collectionId}
-    folderId={explorer.id}
-    {workspaceId}
-    folder={explorer}
-    closePopup={handleFolderPopUp}
-  />
-{/if}
+<ModalWrapperV1
+  title={"Delete Folder?"}
+  type={"danger"}
+  isOpen={isFolderPopup}
+  handleModalState={handleFolderPopUp}
+>
+  <div style="font-size: 14px;" class="text-lightGray mb-1">
+    <p>
+      Are you sure you want to delete this Folder? Everything in <span
+        style="font-weight:700;"
+        class="text-whiteColor">"{explorer.name}"</span
+      >
+      will be removed.
+    </p>
+  </div>
+  <div class="d-flex gap-3" style="font-size:12px">
+    <div class="d-flex gap-1">
+      <span class="text-plusButton">{requestCount}</span>
+      <p>API Requests</p>
+    </div>
+  </div>
+  <div
+    class="d-flex align-items-center justify-content-end gap-3 mt-1 mb-0 rounded"
+    style="font-size: 16px;"
+  >
+    <CustomButton
+      disable={deleteLoader}
+      text={"Cancel"}
+      fontSize={14}
+      type={"dark"}
+      loader={false}
+      onClick={() => {
+        handleFolderPopUp(false);
+      }}
+    />
+
+    <CustomButton
+      disable={deleteLoader}
+      text={"Delete"}
+      fontSize={14}
+      type={"danger"}
+      loader={deleteLoader}
+      onClick={() => {
+        handleDelete();
+      }}
+    />
+  </div></ModalWrapperV1
+>
+<!-- {/if} -->
 
 <svelte:window
   on:click={closeRightClickContextMenu}
