+++ src/lib/components/collections/collections-list/Folder.svelte
@@ -14,7 +14,6 @@
   import Spinner from "$lib/components/Transition/Spinner.svelte";
   import { selectMethodsStore } from "$lib/store/methods";
   import { onDestroy } from "svelte";
-  import CollectionPopup from "$lib/components/Modal/CollectionPopup.svelte";
   import type { Path } from "$lib/utils/interfaces/request.interface";
   import { handleCollectionClick } from "$lib/utils/helpers/handle-clicks.helper";
   import { generateSampleFolder } from "$lib/utils/sample/folder.sample";
@@ -25,6 +24,9 @@
   import { isApiCreatedFirstTime } from "$lib/store/request-response-section";
   import folderIcon from "$lib/assets/create_folder.svg";
   import requestIcon from "$lib/assets/create_request.svg";
+  import ModalWrapperV1 from "$lib/components/Modal/ModalWrapperV1.svelte";
+  import { notifications } from "$lib/utils/notifications";
+  import CustomButton from "$lib/components/buttons/CustomButton.svelte";
 
   export let title: string;
   export let collection: any;
@@ -276,16 +278,65 @@
       disabled: false,
     },
   ];
+  let requestCount: number = 0;
+  let folderCount: number = 0;
+  let deleteIds: string[] = [];
   $: {
     if (activePath) {
       if (activePath.collectionId === collection.id) {
         visibility = true;
       }
     }
+    if (collection) {
+      deleteIds.length = 0;
+      requestCount = 0;
+      folderCount = 0;
+      collection.items.forEach((item) => {
+        if (item.type === ItemType.FOLDER) {
+          deleteIds.push(item.id);
+          folderCount++;
+          requestCount += item.items.length;
+          for (let i = 0; i < item.items.length; i++) {
+            deleteIds.push(item.items[i].id);
+          }
+        }
+        if (item.type === ItemType.REQUEST) {
+          requestCount++;
+          deleteIds.push(item.id);
+        }
+      });
+      deleteIds.push(collectionId);
+    }
   }
+
+  // delete collection
+
+  let deleteLoader: boolean = false;
+  const handleDelete = async () => {
+    deleteLoader = true;
+    const response = await collectionService.deleteCollection(
+      currentWorkspaceId,
+      collectionId,
+    );
+
+    if (response.isSuccessful) {
+      collectionsMethods.deleteCollection(collectionId);
+      collectionsMethods.deleteCollectioninWorkspace(
+        currentWorkspaceId,
+        collectionId,
+      );
+      handleCollectionPopUp(false);
+      notifications.success(`"${collection.name}" Collection deleted.`);
+      collectionsMethods.removeMultipleTabs(deleteIds);
+      deleteLoader = false;
+    } else {
+      notifications.error("Failed to delete the Collection.");
+      deleteLoader = false;
+    }
+  };
 </script>
 
-{#if isCollectionPopup}
+<!-- {#if isCollectionPopup}
   <CollectionPopup
     {collectionsMethods}
     {collection}
@@ -293,7 +344,60 @@
     workspaceId={currentWorkspaceId}
     closePopup={handleCollectionPopUp}
   />
-{/if}
+{/if} -->
+
+<ModalWrapperV1
+  title={"Delete Collection?"}
+  type={"danger"}
+  isOpen={isCollectionPopup}
+  handleModalState={handleCollectionPopUp}
+>
+  <div style="font-size: 14px;" class="text-lightGray mb-1">
+    <p>
+      Are you sure you want to delete this Collection? Everything in <span
+        style="font-weight:700;"
+        class="text-whiteColor">"{collection.name}"</span
+      >
+      will be removed.
+    </p>
+  </div>
+  <div class="d-flex gap-3" style="font-size:12px">
+    <div class="d-flex gap-1">
+      <span class="text-plusButton">{requestCount}</span>
+      <p>API Requests</p>
+    </div>
+    <div class="d-flex gap-1">
+      <span class="text-plusButton">{folderCount}</span>
+      <p>Folder</p>
+    </div>
+  </div>
+  <div
+    class="d-flex align-items-center justify-content-end gap-3 mt-1 mb-0 rounded"
+    style="font-size: 16px;"
+  >
+    <CustomButton
+      disable={deleteLoader}
+      text={"Cancel"}
+      fontSize={14}
+      type={"dark"}
+      loader={false}
+      onClick={() => {
+        handleCollectionPopUp(false);
+      }}
+    />
+
+    <CustomButton
+      disable={deleteLoader}
+      text={"Delete"}
+      fontSize={14}
+      type={"danger"}
+      loader={deleteLoader}
+      onClick={() => {
+        handleDelete();
+      }}
+    />
+  </div></ModalWrapperV1
+>
 
 {#if showMenu}
   <nav style="position: fixed; top:{pos.y}px; left:{pos.x}px; z-index:4;">
